refactor(export): split lib.rs inline tests out to lib_tests.rs for the module-size ratchet
The two serde derive lines on HbjsonStats pushed lib.rs from 399 to 401
lines, one over the 400-line ratchet. The allowlist policy is shrink or
split, so the inline #[cfg(test)] mod tests moves verbatim to
lib_tests.rs (ratchet-exempt via the _tests.rs suffix), the same pattern
as units.rs/units_tests.rs and step_text.rs/step_text_tests.rs.
lib.rs is now 251 lines; no allowlist row added.
